fix(fatiga): castear el parámetro fraccionario antes de la aritmética SQL
Bug introducido en el commit anterior: "$2 * 60 + $3" sin cast deja que
Postgres infiera $2 como integer (por los literales enteros de al lado),
y falla en silencio (atrapado por el catch best-effort del motor de
reglas) apenas minutos_fatiga es fraccionario. Confirmado en vivo: la
prueba de FATIGA de esta noche no disparaba por esto, no por el pulso.

+++ services/backend-api/src/repositories/medicion.repository.js
@@ -101,7 +101,7 @@ const listarVentanaReciente = async (idSeudonimo, minutos) => {
   const query = `
     SELECT * FROM medicion
     WHERE id_seudonimo = $1
-      AND fecha_hora >= now() - (($2 * 60 + $3) || ' seconds')::interval
+      AND fecha_hora >= now() - (($2::numeric * 60 + $3) || ' seconds')::interval
     ORDER BY fecha_hora ASC;
   `;
   const res = await db.query(query, [idSeudonimo, minutos, MARGEN_VENTANA_SEGUNDOS]);
